requirements.txt
numpy>=1.24.0
numba>=0.58.0
pillow>=10.0.0
tqdm>=4.65.0
